refactor: tailoring adapter folder estructure
