+++ src/main/java/pauloh/main/adapter/input/dto/loan/LoanedBooksWithUsersDto.java
@@ -1,4 +1,4 @@
-package pauloh.main.adapter.dto.loan;
+package pauloh.main.adapter.input.dto.loan;
 
 public record LoanedBooksWithUsersDto(String bookTitle, String bookAuthor, String userName, String userEmail) {
