+++ src/main/java/pauloh/main/adapter/input/dto/user/CreateUserDto.java
@@ -1,4 +1,4 @@
-package pauloh.main.adapter.dto.user;
+package pauloh.main.adapter.input.dto.user;
 
 public record CreateUserDto(String name, String email) {
