+++ src/main/java/pauloh/main/adapter/input/dto/user/UserResponseDTO.java
@@ -1,4 +1,4 @@
-package pauloh.main.adapter.dto.user;
+package pauloh.main.adapter.input.dto.user;
 
 import java.util.UUID;
