+++ src/main/java/pauloh/main/adapter/output/repository/entity/UserEntity.java
@@ -1,0 +1,5 @@
+package pauloh.main.adapter.output.repository.entity;
+
+public class UserEntity {
+  // TODO
+}
