--- src/main/java/pauloh/main/adapter/repository/entity/UserEntity.java
@@ -1,5 +1,0 @@
-package pauloh.main.adapter.repository.entity;
-
-public class UserEntity {
-  // TODO
-}
